requirements.txt
streamlit
pdfplumber
pandas
openpyxl
requests
python-dotenv
